selectpage 완료 및 showcardlist 마크업 완료
