--- src/components/BackButton.jsx
@@ -1,22 +1,0 @@
-function BackButton() {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="88"
-      height="88"
-      viewBox="0 0 88 88"
-      fill="none"
-    >
-      <path
-        d="M33.9471 80.0019C32.4269 78.826 30.7835 77.8602 29.1786 76.8394C26.2691 74.987 23.3244 73.1951 20.2774 71.5737C19.1972 70.9995 18.1071 70.4451 17.0225 69.8797C15.0854 68.8699 13.2044 67.7666 11.4004 66.5313C10.5017 65.9164 9.53042 65.4203 8.61302 64.8362C8.05092 64.4787 7.50972 64.0926 7.07082 63.5789C6.88272 63.3589 6.79472 63.1191 6.81892 62.8298C6.83872 62.5867 6.77162 62.3469 6.79032 62.1027C6.80462 61.9212 6.83872 61.7463 6.92782 61.589C7.23802 61.0412 7.56582 60.5088 7.95962 60.0116C8.25112 59.6442 8.50522 59.2438 8.79672 58.872C9.66682 57.7643 10.5479 56.6643 11.3861 55.5313C12.2441 54.3719 13.2121 53.3082 14.1328 52.2038C14.6465 51.5867 15.2196 51.0246 15.7432 50.4185C16.3438 49.7244 16.8234 48.939 17.3767 48.2097C18.0862 47.2747 18.8177 46.3562 19.5206 45.4168C20.2191 44.484 20.8417 43.4951 21.4995 42.5315C21.637 42.3302 21.7734 42.1278 21.9274 41.9397C23.078 40.5427 23.914 38.9301 25.003 37.4913C26.6244 35.3496 27.9939 33.0396 29.4756 30.8044C29.5658 30.668 29.6714 30.5404 29.7737 30.4117C30.085 30.0201 30.448 29.832 30.9859 29.9233C31.4952 30.0091 31.8164 30.2687 31.9363 30.7263C32.2047 31.757 32.4368 32.7965 32.4962 33.869C32.5446 34.7479 32.7019 35.6158 32.835 36.4859C33.0176 37.6805 33.1892 38.8784 33.3817 40.0719C33.5863 41.3358 33.77 42.603 33.9064 43.8768C33.9746 44.517 34.012 45.1594 34.0648 45.8018C34.0714 45.8876 34.0857 45.9723 34.1011 46.057C34.1693 46.4475 34.1704 46.4442 34.5686 46.4926C35.2451 46.5751 35.9216 46.5905 36.6036 46.5432C37.1888 46.5025 37.7773 46.5069 38.3647 46.508C39.0885 46.5091 39.8101 46.4695 40.5317 46.42C40.9178 46.3936 41.305 46.3628 41.69 46.3716C42.8571 46.3969 44.0011 46.2022 45.1495 46.0482C45.76 45.9668 46.3727 45.9074 46.9876 45.8799C47.7785 45.8447 48.5419 45.6434 49.313 45.4927C51.0895 45.1451 52.8594 44.7689 54.6183 44.3421C55.7183 44.0759 56.8051 43.7613 57.871 43.3796C58.3748 43.1992 58.861 42.9627 59.3395 42.7207C59.6629 42.5579 59.8499 42.2433 60.0083 41.9309C60.5341 40.8925 60.7904 39.7694 60.9972 38.6342C61.1314 37.8961 61.2942 37.1613 61.3404 36.4089C61.3437 36.3594 61.3525 36.3099 61.3547 36.2593C61.446 32.78 61.4295 29.3007 61.3833 25.8225C61.3679 24.6499 61.2458 23.4806 61.1457 22.3113C61.0808 21.5556 61.0489 20.7977 61.0522 20.0376C61.0588 18.6395 60.9158 17.2513 60.7695 15.8631C60.6287 14.5244 60.3746 13.2 60.2569 11.8613C60.1843 11.0308 60.17 10.1893 60.2866 9.35548C60.335 9.00678 60.5341 8.78678 60.8641 8.67788C61.1248 8.59208 61.3965 8.57668 61.6682 8.55798C64.6789 8.35228 67.694 8.40398 70.708 8.38528C71.4824 8.38088 72.2502 8.43038 73.0213 8.49528C73.9651 8.57448 74.9067 8.39958 75.8505 8.38858C76.8372 8.37648 77.8195 8.51618 78.8073 8.48648C79.1494 8.47658 79.4277 8.74278 79.4728 9.07718C79.5245 9.45778 79.5751 9.83288 79.5025 10.2267C79.4497 10.5138 79.5234 10.824 79.5454 11.1232C79.6389 12.3937 79.6081 13.6708 79.7665 14.9391C79.8204 15.3725 79.8457 15.8081 79.9634 16.2316C80.1845 17.0236 80.1834 17.8453 80.2714 18.6549C80.3539 19.4095 80.4419 20.163 80.5321 20.9154C80.5871 21.3708 80.5596 21.8306 80.6091 22.2871C80.7818 23.859 80.8423 25.4397 80.9633 27.016C81.0623 28.3162 81.0007 29.6208 81.0117 30.9243C81.0205 31.9836 81.0205 33.044 81.0161 34.1033C81.0106 35.1912 80.9259 36.2769 80.872 37.3626C80.8082 38.643 80.6542 39.9157 80.5365 41.1906C80.4078 42.5942 80.1746 43.9868 79.9194 45.3728C79.728 46.4134 79.4354 47.4331 79.2099 48.466C79.0196 49.3339 78.8128 50.1963 78.5081 51.0312C78.2463 51.7495 77.9691 52.4623 77.6215 53.1476C77.3949 53.5931 77.209 54.0595 76.9769 54.5039C76.1365 56.1143 75.0134 57.4992 73.678 58.7268C72.0709 60.2052 70.4572 61.6759 68.6433 62.9035C68.2143 63.1939 67.7468 63.4128 67.2749 63.6174C66.7359 63.8506 66.2156 64.1047 65.7272 64.4391C65.0298 64.9176 64.2235 65.1222 63.3963 65.2498C62.8375 65.3367 62.2842 65.4555 61.7232 65.5248C60.4527 65.6821 59.1756 65.7976 57.9073 65.9681C56.2034 66.1969 54.4907 66.3157 52.7769 66.4444C51.8397 66.5148 50.9047 66.5434 49.9675 66.5456C49.0061 66.5467 48.0535 66.6622 47.0954 66.7205C45.7589 66.803 44.4202 66.8305 43.0815 66.8261C41.756 66.8217 40.436 66.9383 39.1138 67.012C38.6485 67.0384 38.1832 67.0164 37.7179 66.9955C37.5738 66.9889 37.5265 66.9988 37.5419 67.1759C37.6596 68.5212 37.8972 69.8533 37.939 71.2107C37.9874 72.7903 38.1381 74.3677 38.2547 75.9451C38.3174 76.7943 38.3933 77.6424 38.4824 78.4894C38.5484 79.1153 38.6782 79.7324 38.7706 80.3539C38.8542 80.9182 38.8641 81.4913 38.7574 82.0567C38.6694 82.5209 38.2118 82.8047 37.7311 82.7002C37.5067 82.6518 37.2878 82.5814 37.0821 82.4791C36.2692 82.0754 35.5168 81.5914 34.9217 80.8885C34.6434 80.5552 34.3299 80.2505 33.9471 80.0019Z"
-        fill="black"
-      />
-      <path
-        d="M76.4698 21.5457C76.6051 23.4025 76.5974 25.2428 76.6216 27.0831C76.6491 29.1302 76.7481 31.1762 76.8493 33.2211C76.9923 36.1031 77.1089 38.9884 77.0275 41.8737C76.9604 44.2772 76.5787 46.6279 75.6767 48.8785C74.9595 50.6649 73.9607 52.2687 72.7133 53.7284C71.3735 55.2959 69.8896 56.7094 68.2583 57.97C66.9471 58.9831 65.538 59.8301 63.9848 60.4076C62.8771 60.819 61.7265 61.0929 60.5737 61.3569C58.5695 61.8156 56.5455 62.1511 54.5039 62.3986C53.1927 62.557 51.8771 62.5933 50.5604 62.6065C48.3549 62.6274 46.1494 62.6087 43.9461 62.7352C42.3456 62.8276 40.7451 62.9354 39.1435 63.0058C37.2768 63.0883 35.409 63.1367 33.5423 63.2005C33.2057 63.2126 33.2079 63.2236 33.1925 63.5613C33.1265 64.9759 33.3586 66.3542 33.6556 67.7292C33.8536 68.6444 33.9438 69.5805 34.1407 70.4979C34.4619 71.9917 34.6027 73.5086 34.7061 75.031C34.7567 75.7801 34.848 76.5259 34.9217 77.2893C34.8095 77.2651 34.7512 77.1815 34.6852 77.1155C33.0341 75.4589 31.2356 73.975 29.3513 72.5945C28.1149 71.6881 26.7905 70.9192 25.4837 70.1195C22.4796 68.2814 19.4084 66.5566 16.4054 64.7174C15.301 64.0409 14.2351 63.3017 13.0889 62.6923C12.4256 62.3392 11.7546 61.9993 11.0506 61.7331C10.8691 61.6649 10.868 61.6022 10.9747 61.4526C11.7898 60.3097 12.6852 59.2295 13.5443 58.1207C14.784 56.5191 15.9896 54.8911 17.2029 53.2686C18.8232 51.1016 20.5414 49.0105 22.1771 46.8545C24.1461 44.2596 26.1162 41.6658 27.7629 38.8465C28.2546 38.005 28.6891 37.1349 29.0939 36.2483C29.1225 36.1856 29.1346 36.0822 29.2138 36.0888C29.3062 36.0965 29.282 36.201 29.2853 36.267C29.3876 38.1535 29.7759 39.9971 30.1378 41.844C30.3941 43.1519 30.536 44.4763 30.6768 45.7996C30.8088 47.0371 30.9397 48.2746 31.0739 49.5121C31.0937 49.6892 31.1377 49.8641 31.1575 50.0412C31.174 50.1974 31.2576 50.2524 31.4028 50.2359C34.2221 49.9257 37.0524 50.149 39.8761 50.0731C41.6757 50.0247 43.461 49.8773 45.2452 49.6419C47.7741 49.3086 50.2788 48.8444 52.7747 48.3307C54.8174 47.9105 56.8381 47.4012 58.7818 46.6334C61.0071 45.7534 62.6494 44.242 63.6768 42.0585C64.6536 39.9828 65.2366 37.8114 65.4841 35.5333C65.7173 33.3872 65.637 31.2433 65.5358 29.0983C65.4357 26.972 65.1134 24.8677 64.8637 22.7568C64.6008 20.5315 64.3885 18.3018 64.218 16.0666C64.0981 14.4969 63.9617 12.9272 63.5041 11.407C63.4601 11.2618 63.5239 11.2332 63.6537 11.2409C64.7262 11.2992 65.7965 11.3784 66.8602 11.528C68.4057 11.7458 69.96 11.88 71.5165 12.001C73.051 12.1209 74.5877 12.2067 76.1189 12.3673C76.219 12.3783 76.2608 12.4091 76.2564 12.5092C76.2509 12.6379 76.2542 12.7666 76.252 12.8953C76.2146 15.2075 76.1838 17.5208 76.3576 19.8297C76.4005 20.3973 76.4324 20.9616 76.4698 21.5457Z"
-        fill="#FFC63E"
-      />
-    </svg>
-  );
-}
-
-export default BackButton;
