+++ src/components/Envelope.jsx
@@ -1,0 +1,9 @@
+function Envelope() {
+  return (
+    <>
+      <img src="/envelope_home.svg" alt="편지" />
+    </>
+  );
+}
+
+export default Envelope;
