+++ src/layout/Header.jsx
@@ -1,17 +1,18 @@
-import Lion from '@/components/Lion';
-import Arrow from '@/components/Arrow';
+import Lion from "@/components/Lion";
+import Arrow from "@/components/Arrow";
 
-function Header() {
+function Header({ text }) {
   return (
-    <header className="flex bg-lionGray w-[1280px] h-[100px] items-center">
+    <header className="flex bg-lionGray w-full h-[100px] items-center">
       <div className="my-3.5 ml-10 mr-4">
         <Lion width={85} height={85} />
       </div>
       <div className=" flex w-[870px] h-[50px] text-[40px] font-bold items-end">
-        <h1 className="mr-4">움직이는 우표를 골라보세요!</h1>
-        <p className="text-xl">(feat.GIFHY)</p>
+        <h1 className="mr-4">{text}</h1>
       </div>
-      <Arrow />
+      <div>
+        <Arrow />
+      </div>
     </header>
   );
 }
